chore: Create a basic string manipulation
Now the basic code that does basic find and replace has been finalized.

The following have yet to be implemented:

- No type checks are done, anything is printed.
- Only two types are supported
- No error is raised on incorrect values.

Other buitins have been moved to another file to create room to extend the `Andika` function.
This is intended to other builtins.
