+++ builtins/andika.go
@@ -1,11 +1,8 @@
 package builtins
 
 import (
-	"bufio"
 	"fmt"
 	"github.com/AvicennaJr/Nuru/object"
-	"io"
-	"os"
 	"strings"
 )
 
@@ -20,6 +17,15 @@ func Andika(args ...object.Object) object.Object {
 			}
 			arr = append(arr, arg.Inspect())
 		}
+
+		passedType := Aina(args[0]).Inspect()
+		if passedType == "NENO" {
+				nStr := arr[1:]
+				drrr := formatString(arr[0], nStr)
+				fmt.Println(drrr)
+				return nil
+		}
+
 		str := strings.Join(arr, " ")
 		fmt.Println(str)
 	}
@@ -30,48 +36,24 @@ func newError(format string, a ...interface{}) *object.Error {
 	return &object.Error{Message: fmt.Sprintf(format, a...)}
 }
 
-func Aina(args ...object.Object) object.Object {
-	if len(args) != 1 {
-		return newError("Samahani, tunahitaji hoja 1, wewe umeweka %d", len(args))
+func formatString(fmt string, vals []string) string {
+	var newStr strings.Builder
+	pos := 0
+	for i := 0; i < len(fmt); i++ {
+		if fmt[i] == '%' && i+1 < len(fmt) {
+			switch fmt[i+1] {
+			case 's', 'd':
+				if pos < len(vals) {
+					newStr.WriteString(vals[pos])
+					pos++
+					i++
+					continue
+				}
+			}
+		} else {
+			newStr.WriteByte(fmt[i])
+		}
 	}
-
-	return &object.String{Value: string(args[0].Type())}
+	return newStr.String()
 }
 
-func Fungua(args ...object.Object) object.Object {
-
-	if len(args) != 1 {
-		return newError("Samahani, tunahitaji hoja 1, wewe umeweka %d", len(args))
-	}
-	filename := args[0].(*object.String).Value
-
-	file, err := os.ReadFile(filename)
-	if err != nil {
-		return &object.Error{Message: "Tumeshindwa kusoma faili"}
-	}
-	return &object.File{Filename: filename, Content: string(file)}
-}
-
-func Jaza(args ...object.Object) object.Object {
-
-	if len(args) > 1 {
-		return newError("Samahani, kiendesha hiki kinapokea hoja 0 au 1, wewe umeweka %d", len(args))
-	}
-
-	if len(args) > 0 && args[0].Type() != object.STRING_OBJ {
-		return newError(fmt.Sprintf(`Tafadhali tumia alama ya nukuu: "%s"`, args[0].Inspect()))
-	}
-	if len(args) == 1 {
-		prompt := args[0].(*object.String).Value
-		fmt.Fprint(os.Stdout, prompt)
-	}
-
-	buffer := bufio.NewReader(os.Stdin)
-
-	line, _, err := buffer.ReadLine()
-	if err != nil && err != io.EOF {
-		return newError("Nimeshindwa kusoma uliyo yajaza")
-	}
-
-	return &object.String{Value: string(line)}
-}
